# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PersonaFit API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/analyses": {
      "get": {
        "tags": [
          "analyses"
        ],
        "summary": "List Endpoint",
        "description": "분석 이력 요약 리스트 (최신순).",
        "operationId": "list_endpoint_api_analyses_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysesListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "analyses"
        ],
        "summary": "Delete All Endpoint",
        "description": "모든 분석 + 시뮬레이션 일괄 삭제 (되돌릴 수 없음).",
        "operationId": "delete_all_endpoint_api_analyses_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete All Endpoint Api Analyses Delete"
                }
              }
            }
          }
        }
      }
    },
    "/api/analyses/{analysis_id}": {
      "get": {
        "tags": [
          "analyses"
        ],
        "summary": "Detail Endpoint",
        "description": "분석 이력 단건 전체 (4섹션 다시 렌더링용).\n\n응답은 AnalyzeResponse와 동일 구조 + 추가 메타(id, created_at, product_text 일부)\n+ simulations(이 분석에 묶인 과거 시뮬레이션 전체, 최신순).",
        "operationId": "detail_endpoint_api_analyses__analysis_id__get",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Detail Endpoint Api Analyses  Analysis Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "analyses"
        ],
        "summary": "Delete Endpoint",
        "description": "분석 단건 + 연관 시뮬레이션 삭제.",
        "operationId": "delete_endpoint_api_analyses__analysis_id__delete",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Endpoint Api Analyses  Analysis Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalysesListResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/AnalysisSummary"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "total",
          "items"
        ],
        "title": "AnalysesListResponse"
      },
      "AnalysisSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "key_benefits": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Key Benefits"
          },
          "max_score": {
            "type": "number",
            "title": "Max Score"
          },
          "top_persona_count": {
            "type": "integer",
            "title": "Top Persona Count"
          },
          "top_province": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top Province"
          },
          "top_province_count": {
            "type": "integer",
            "title": "Top Province Count",
            "default": 0
          },
          "total_ms": {
            "type": "integer",
            "title": "Total Ms"
          },
          "simulation_count": {
            "type": "integer",
            "title": "Simulation Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at",
          "summary",
          "key_benefits",
          "max_score",
          "top_persona_count",
          "total_ms"
        ],
        "title": "AnalysisSummary"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}